Move epyccel_test class to tests.epyccel.utilities. Improve its docstring
