+++ tests/epyccel/test_compare_expressions.py
@@ -1,22 +1,6 @@
 # pylint: disable=missing-function-docstring, missing-module-docstring/
-import numpy as np
-from pyccel.epyccel import epyccel
 from pyccel.decorators import types
-
-#==============================================================================
-class epyccel_test:
-    """
-    Class to pyccelize module or function then compare different results
-    This avoids the need to pyccelize the object multiple times
-    """
-    def __init__(self, f, lang='fortran'):
-        self._f  = f
-        self._f2 = epyccel(f, language=lang)
-
-    def compare_epyccel(self, *args):
-        out1 = self._f(*args)
-        out2 = self._f2(*args)
-        assert np.equal(out1, out2 )
+from .utilities import epyccel_test
 
 #==============================================================================
 @types('int, int, int')
